* Fix segfault hash_xorpair on big endian * Fix typo in help msg (value -> data) * Graph now are more deeper!

+++ src/hasher/hash.h
@@ -1,5 +1,5 @@
 int hash_par(unsigned char *buffer, u64 len);
-u16 hash_xorpair(const u16 *b, int len);
+u16 hash_xorpair(const u8 *b, u64 len);
 u8  hash_xor(const u8 *b, u64 len);
 u8  hash_mod255(const u8 *b, u64 len);
 u32 hash_wrt54gv5v6(const u8 *pStart, u64 len);
